Refactor inspector stuff out of timeline into its own files
